face_recognition/requirements.txt
Flask~=3.0.3
SQLAlchemy
psycopg2-binary
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
flask_cors~=3.0.10
dlib~=19.24.0
scikit-learn~=1.5.0
Pillow~=10.2.0
torch~=2.2.2
facenet_pytorch
joblib
numpy~=1.26.4
opencv-python~=4.10.0.82
tensorflow
scikit-image~=0.24.0